refactor: 온보딩 관련 쿼리 구조 변경
