+++ apps/client/src/shared/apis/onboard/artist-related-queries.ts
@@ -1,13 +1,12 @@
 import { queryOptions } from '@tanstack/react-query';
 
+import { get } from '@shared/apis/config/instance';
+import { END_POINT } from '@shared/constants/api';
 import { ARTIST_RELATED_QUERY_KEY } from '@shared/constants/query-key';
+import { BaseResponse } from '@shared/types/api';
+import { onboardResponse } from '@shared/types/onboard-response';
 
-import {
-  getArtistRelatedArtist,
-  getArtistRelatedKeyword,
-} from './artist-related';
-
-export const ARTIST_RELATED_QUERY_OPTION = {
+export const ARTIST_RELATED_QUERY_OPTIONS = {
   ALL: () => queryOptions({ queryKey: ARTIST_RELATED_QUERY_KEY.ALL }),
   ARTIST: (artistId: string, limit: number) => ({
     queryKey: ARTIST_RELATED_QUERY_KEY.RELATED_ARTIST(artistId),
@@ -18,4 +17,24 @@ export const ARTIST_RELATED_QUERY_OPTION = {
     queryKey: ARTIST_RELATED_QUERY_KEY.RELATED_KEYWORD(keyword),
     queryFn: () => getArtistRelatedKeyword(keyword, limit),
   }),
+};
+
+export const getArtistRelatedArtist = async (
+  artistId: string,
+  limit: number,
+): Promise<BaseResponse<onboardResponse>> => {
+  const response = await get<BaseResponse<onboardResponse>>(
+    END_POINT.GET_ARTIST_RELATED_ARTIST(artistId, limit),
+  );
+  return response;
+};
+
+export const getArtistRelatedKeyword = async (
+  keyword: string,
+  limit: number,
+): Promise<onboardResponse> => {
+  const response = await get<BaseResponse<onboardResponse>>(
+    END_POINT.GET_ARTIST_RELATED_KEYWORDS(keyword, limit),
+  );
+  return response.data;
 };
